add a move widget class and crude available moves preview
A Move class was created that inherits from QWidget. This allows for easy drawing and mouse interaction.

The Move class in currently in general_TA_assembly.py for testing purposes only. I plan to move it out
once initial testing and functionality are complete.

Currently for each draw_tiles call, a move widget will be created for every element in the
validMoves list.